pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "anthropic>=0.86.0",
    "instructor>=1.14.5",
    "langchain-anthropic>=1.4.0",
    "langchain-neo4j>=0.9.0",
    "neo4j>=6.1.0",
    "plotly>=5.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
